test(find): sui-1669 - initial changes for parallelisation of tests
