+++ Apps/Find/tests/SUI.Find.E2ETests/StartANewSearchTestsV1.cs
@@ -1,0 +1,12 @@
+using Xunit.Abstractions;
+
+namespace SUI.Find.E2ETests;
+
+[Collection("E2E-V1")]
+public class StartANewSearchTestsV1 : StartANewSearchTestsBase
+{
+    protected override bool UsePolling => false;
+
+    public StartANewSearchTestsV1(FunctionTestFixture fixture, ITestOutputHelper output)
+        : base(fixture, output) { }
+}
